web_news/requirements.txt
Django
djangorestframework
django-cors-headers
sentry-sdk
requests
gunicorn
loguru
python-dotenv
drf-spectacular
validate-docbr
flake8
black
bandit
ipython
isort
mock
django-debug-toolbar
requests-mock
pylint
pylint-django
pytest
pytest-django
pytest-pylint
pytest-black
pytest-coverage
pytest-mock
coverage
codacy-coverage
pre-commit
factory-boy
asgiref
certifi
cffi
chardet
cryptography
defusedxml
django-templated-mail
idna
pycparser
PyJWT
python3-openid
pytz
six
sqlparse
urllib3
bcrypt
arrow
psycopg2
psycopg2-binary
pypdf
docx2txt
